Changed library interface. Fixed namespace problems with instruction lookup in library.
	* configure: Regenerated.
	* configure.ac: Version changed to current date.
	* cpu/or1k/opcode/or32.h <or1ksim_build_automata>: Renamed from
	build_automata.
	<l_none, num_opcodes, insn_index>: Deleted.
	<or1ksim_op_start>: Renamed from op_start.
	<or1ksim_automata>: Renamed from automata.
	<or1ksim_ti>: Renamed from ti.
	<or1ksim_or32_opcodes>: Renamed from or32_opcodes.
	<or1ksim_disassembled>: Renamed from disassembled.
	<or1ksim_insn_len>: Renamed from insn_len.
	<or1ksim_insn_name>: Renamed from insn_name.
	<or1ksim_destruct_automata>: Renamed from destruct_automata.
	<or1ksim_insn_decode>: Renamed from insn_decode.
	<or1ksim_disassemble_insn>: Renamed from disassemble_insn.
	<or1ksim_disassemble_index>: Renamed from disassemble_index.
	<or1ksim_extend_imm>: Renamed from extend_imm.
	<or1ksim_or32_extract>: Renamed from or32_extract
	* cpu/or32/or32.c, cpu/or32/execute.c, cpu/or32/generate.c,
	* cpu/common/stats.c, cpu/common/abstract.c, cpu/common/parse.c,
	* cpu/or1k/opcode/or32.h, cuc/load.c, cuc/cuc.c,
	* support/dumpverilog.c, toplevel-support.c: Renaming
	corresponding to changes in cpu/or1k/opcode/or32.h.
	* cpu/or32/execute-fp.h: Deleted
	* cpu/or32/generate.c <include_strings>: Remove reference to
	execute-fp.h
	* cpu/or32/execute.c <host_fp_rm>: Declared static.
	(fp_set_flags_restore_host_rm, fp_set_or1k_rm): Declared static,
	forward declaration removed.
	* or1ksim.h (or1ksim_read_mem, or1ksim_write_mem): addr arg
	changed to unsigned long int.
	(or1ksim_read_spr): sprval_ptr arg changed to unsigned long int *.
	(or1ksim_write_spr): sprval arg changed to unsigned long int.
	(or1ksim_read_reg): regval_ptr arg changed to unsigned long int *.
	(or1ksim_write_reg): regval arg changed to unsigned long int.
	* libtoplevel.c (or1ksim_read_mem, or1ksim_write_mem): addr arg
	changed to unsigned long int.
	(or1ksim_read_spr): sprval_ptr arg changed to unsigned long int *.
	(or1ksim_write_spr): sprval arg changed to unsigned long int.
	(or1ksim_read_reg): regval_ptr arg changed to unsigned long int *.
	(or1ksim_write_reg): regval arg changed to unsigned long int.

+++ libtoplevel.c
@@ -525,9 +525,9 @@ or1ksim_jtag_shift_dr (unsigned char *jreg,
    @return  Number of bytes read, or zero if error.                          */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_read_mem (unsigned int   addr,
-		  unsigned char *buf,
-		  int            len)
+or1ksim_read_mem (unsigned long int  addr,
+		  unsigned char     *buf,
+		  int                len)
 {
   int             off;			/* Offset into the memory */
 
@@ -562,9 +562,9 @@ or1ksim_read_mem (unsigned int   addr,
    @return  Number of bytes written, or zero if error.                       */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_write_mem (unsigned int   addr,
-		   unsigned char *buf,
-		   int            len)
+or1ksim_write_mem (unsigned long int  addr,
+		   unsigned char     *buf,
+		   int                len)
 {
   int             off;			/* Offset into the memory */
 
@@ -598,8 +598,8 @@ or1ksim_write_mem (unsigned int   addr,
    @return  Non-zero (TRUE) on success, zero (FALSE) otherwise.              */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_read_spr (int            sprnum,
-		  unsigned int  *sprval_ptr)
+or1ksim_read_spr (int                sprnum,
+		  unsigned long int *sprval_ptr)
 {
   /* SPR numbers are up to 16 bits long */
   if ((unsigned int) sprnum <= 0xffff)
@@ -623,8 +623,8 @@ or1ksim_read_spr (int            sprnum,
    @return  Non-zero (TRUE) on success, zero (FALSE) otherwise.              */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_write_spr (int           sprnum,
-		   unsigned int  sprval)
+or1ksim_write_spr (int                sprnum,
+		   unsigned long int  sprval)
 {
   /* SPR numbers are up to 16 bits long */
   if ((unsigned int) sprnum <= 0xffff)
@@ -653,8 +653,8 @@ or1ksim_write_spr (int           sprnum,
    @return  Non-zero (TRUE) on success, zero (FALSE) otherwise.              */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_read_reg (int            regnum,
-		  unsigned int  *regval_ptr)
+or1ksim_read_reg (int                 regnum,
+		  unsigned long int  *regval_ptr)
 {
   /* GPR's */
   if (regnum < MAX_GPRS)
@@ -689,8 +689,8 @@ or1ksim_read_reg (int            regnum,
    @return  Non-zero (TRUE) on success, zero (FALSE) otherwise.              */
 /*---------------------------------------------------------------------------*/
 int
-or1ksim_write_reg (int           regnum,
-		   unsigned int  regval)
+or1ksim_write_reg (int                regnum,
+		   unsigned long int  regval)
 {
   /* GPR's */
   if (regnum < MAX_GPRS)
