+++ or1ksim.h
@@ -90,25 +90,25 @@ double  or1ksim_jtag_shift_dr (unsigned char *jreg,
 			       int            num_bits);
 
 /* Access to simulator state */
-int  or1ksim_read_mem (unsigned int   addr,
-		       unsigned char *buf,
-		       int            len);
+int  or1ksim_read_mem (unsigned long int  addr,
+		       unsigned char     *buf,
+		       int                len);
 
-int  or1ksim_write_mem (unsigned int   addr,
-			unsigned char *buf,
-			int            len);
+int  or1ksim_write_mem (unsigned long int  addr,
+			unsigned char     *buf,
+			int                len);
 
-int  or1ksim_read_spr (int            sprnum,
-		       unsigned int  *sprval_ptr);
+int  or1ksim_read_spr (int                 sprnum,
+		       unsigned long int  *sprval_ptr);
 
-int  or1ksim_write_spr (int           sprnum,
-			unsigned int  sprval);
+int  or1ksim_write_spr (int                sprnum,
+			unsigned long int  sprval);
     
-int  or1ksim_read_reg (int           regnum,
-		       unsigned int *regval_ptr);
+int  or1ksim_read_reg (int                regnum,
+		       unsigned long int *regval_ptr);
 
-int  or1ksim_write_reg (int           regnum,
-			unsigned int  regval);
+int  or1ksim_write_reg (int                regnum,
+			unsigned long int  regval);
 
 void  or1ksim_set_stall_state (int  state);
